+++ driven/notifications/adapter.go
@@ -11,7 +11,7 @@ import (
 
 // Adapter implements the Notifications interface
 type Adapter struct {
-	host           string
+	baseURL        string
 	internalAPIKey string
 	appID          string
 	orgID          string
@@ -20,52 +20,45 @@ type Adapter struct {
 // NewNotificationsAdapter creates a new Notifications BB adapter instance
 // NewNotificationsAdapter creates a new notifications BB adapter
 func NewNotificationsAdapter(notificationHost string, internalAPIKey string, appID string, orgID string) *Adapter {
-	return &Adapter{host: notificationHost, internalAPIKey: internalAPIKey, appID: appID, orgID: orgID}
+	return &Adapter{baseURL: notificationHost, internalAPIKey: internalAPIKey, appID: appID, orgID: orgID}
+}
+
+// SendNotification Sends a direct notification trough Notifications BB
+func (a *Adapter) SendNotification(notification model.NotificationMessage) {
+	go a.sendNotification(notification)
 }
 
 // SendNotification sends notification to a user
-func (a *Adapter) SendNotification(recipients []model.NotificationRecipient, topic *string, title string, text string, data map[string]string) error {
-	if len(recipients) > 0 {
-		url := fmt.Sprintf("%s/api/int/v2/message", a.host)
+func (a *Adapter) sendNotification(notification model.NotificationMessage) {
+	if len(notification.Recipients) > 0 && notification.Subject != "" && notification.Body != "" {
+		url := fmt.Sprintf("%s/api/int/message", a.baseURL)
 
-		bodyData := model.NotificationMessage{
-			Priority:   10,
-			Topic:      topic,
-			Recipients: recipients,
-			Subject:    title,
-			Body:       text,
-			Data:       data,
-			AppID:      a.appID,
-			OrgID:      a.orgID,
-		}
-		bodyBytes, err := json.Marshal(bodyData)
+		bodyBytes, err := json.Marshal(notification)
 		if err != nil {
 			log.Printf("error creating notification request - %s", err)
-			return err
+			return
 		}
 
 		client := &http.Client{}
 		req, err := http.NewRequest("POST", url, bytes.NewReader(bodyBytes))
 		if err != nil {
 			log.Printf("error creating load user data request - %s", err)
-			return err
+			return
 		}
 		req.Header.Set("INTERNAL-API-KEY", a.internalAPIKey)
 
 		resp, err := client.Do(req)
 		if err != nil {
 			log.Printf("error loading user data - %s", err)
-			return err
+			return
 		}
 
 		defer resp.Body.Close()
 
 		if resp.StatusCode != 200 {
 			log.Printf("error with response code - %d", resp.StatusCode)
-			return fmt.Errorf("error with response code != 200")
 		}
 	}
-	return nil
 }
 
 // SendMail sends email to a user
@@ -75,7 +68,7 @@ func (a *Adapter) SendMail(toEmail string, subject string, body string) {
 
 func (a *Adapter) sendMail(toEmail string, subject string, body string) error {
 	if len(toEmail) > 0 && len(subject) > 0 && len(body) > 0 {
-		url := fmt.Sprintf("%s/api/int/mail", a.host)
+		url := fmt.Sprintf("%s/api/int/mail", a.baseURL)
 
 		bodyData := map[string]interface{}{
 			"to_mail": toEmail,