+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Improve survey models for BESSI [#40](https://github.com/rokwire/polls-building-block/issues/40)
 
 ## [1.3.0] - 2022-11-10
 ### Added