+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## Added
+### Added
+- Improve survey models for BESSI [#40](https://github.com/rokwire/polls-building-block/issues/40)
+- Send notifications for group polls directly through Groups BB [#30](https://github.com/rokwire/polls-building-block/issues/30)
 - Set appID and orgID for notifications [#26](https://github.com/rokwire/polls-building-block/issues/26)
 
 ## [1.3.0] - 2022-11-10